최신 빌드인지 확인

+++ python_server/fastApiServer.py
@@ -1,12 +1,12 @@
 import asyncio
 import shutil
 import time
+from typing import List
 
 import uvicorn
 from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
-from starlette.background import BackgroundTasks
 from starlette.responses import FileResponse
 import json
 import httpx
@@ -40,24 +40,30 @@ async def startup_event():
     try:
         build_url = await get_build_url()
         build_info = await get_json(build_url)
-        file_list = ["/index.html"] + [
-                "/build/" + path_obj["path"]
-                for path_obj in build_info["tree"]
-            ]
+
+        file_list = [path_obj["path"] for path_obj in build_info["tree"]]
         print(file_list)
-
-        if os.path.exists("output/build"):
-            move_builds_to_old_dir()
-            os.rmdir("output/build")
-
-        os.mkdir("output/build")
-        await asyncio.wait([get_and_write_file(file_name) for file_name in file_list])
-
+        if not await download_latest_build(file_list):
+            print("이미 최신 빌드입니다.")
         duration = time.time() - start
         print("소요시간 %s 초, 빌드 파일을 최신화했습니다." % str(int(duration * 100) / 100))
     except Exception as e:
         print("최신 빌드 로딩에 실패했습니다.")
         print(e)
+
+
+async def download_latest_build(file_list: List[str]) -> bool:
+    if os.path.exists("output/build") and set(file_list) == set(os.listdir("output/build")):
+        return False
+
+    if os.path.exists("output/build"):
+        move_builds_to_old_dir()
+        os.rmdir("output/build")
+
+    os.mkdir("output/build")
+    path_list = ["/index.html"] + ["/build/" + file_name for file_name in file_list]
+    await asyncio.wait([get_and_write_file(path) for path in path_list])
+    return True
 
 
 async def get_build_url():
@@ -95,18 +101,6 @@ async def get_and_write_file(file_name):
     )
 
 
-@app.on_event("shutdown")
-async def shutdown_event():
-    await client.aclose()
-
-
-@app.get("/")
-async def root(background_tasks: BackgroundTasks):
-    if os.path.exists('output/index.html'):
-        return FileResponse('output/index.html')
-    raise HTTPException(status_code=404, detail=f"github 업데이트를 불러오는데 실패했습니다. 백업된 파일이 없습니다.")
-
-
 async def get_content(url: str):
     r = await client.get(url)
     return r.text
@@ -120,6 +114,18 @@ async def get_json(url: str):
 def write_file(file_path: str, content: str):
     with open(file_path, mode="w", encoding="UTF-8") as f:
         f.write(content)
+
+
+@app.on_event("shutdown")
+async def shutdown_event():
+    await client.aclose()
+
+
+@app.get("/")
+async def root():
+    if os.path.exists('output/index.html'):
+        return FileResponse('output/index.html')
+    raise HTTPException(status_code=404, detail=f"저장된 빌드 파일이 없습니다.")
 
 
 app.include_router(api_router)
